small cleanup of path constraint setup which turned into a larger refactor
