+++ dymos/transcriptions/pseudospectral/gauss_lobatto.py
@@ -265,14 +265,9 @@ class GaussLobatto(PseudospectralBase):
                           'collocation_constraint.f_computed:{0}'.format(name),
                           src_indices=src_idxs, flat_src_indices=True)
 
-    def setup_path_constraints(self, phase):
-        gd = self.grid_data
-
-        if phase._path_constraints:
-            path_comp = PathConstraintComp(num_nodes=gd.num_nodes)
-            phase.add_subsystem('path_constraints', subsys=path_comp)
-
     def configure_path_constraints(self, phase):
+        super(GaussLobatto, self).configure_path_constraints(phase)
+
         time_units = phase.time_options['units']
         path_comp = phase._get_subsystem('path_constraints')
 
@@ -287,101 +282,50 @@ class GaussLobatto(PseudospectralBase):
             var_type = phase.classify_var(var)
 
             if var_type == 'time':
-                phase.connect(src_name='time',
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                shape = (1,)
-                units = time_units if con_units is None else con_units
-                linear = True
+                src = 'time'
+                tgt = f'path_constraints.all_values:{con_name}'
 
             elif var_type == 'time_phase':
-                phase.connect(src_name='time_phase',
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                shape = (1,)
-                units = time_units if con_units is None else con_units
-                linear = True
+                src = 'time_phase'
+                tgt = f'path_constraints.all_values:{con_name}'
 
             elif var_type == 'state':
-                phase.connect(src_name='interleave_comp.all_values:states:{0}'.format(var),
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                state_shape = phase.state_options[var]['shape']
-                state_units = phase.state_options[var]['units']
-                shape = state_shape
-                units = state_units if con_units is None else con_units
-                linear = False
+                src = f'interleave_comp.all_values:states:{var}'
+                tgt = f'path_constraints.all_values:{con_name}'
 
             elif var_type in ('indep_control', 'input_control'):
-                constraint_path = 'control_values:{0}'.format(var)
-                phase.connect(src_name=constraint_path,
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                control_shape = phase.control_options[var]['shape']
-                control_units = phase.control_options[var]['units']
-                shape = control_shape
-                units = control_units if con_units is None else con_units
-                linear = True
+                src = f'control_values:{var}'
+                tgt = f'path_constraints.all_values:{con_name}'
 
             elif var_type in ('indep_polynomial_control', 'input_polynomial_control'):
-                constraint_path = 'polynomial_control_values:{0}'.format(var)
-                phase.connect(src_name=constraint_path,
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                control_shape = phase.polynomial_control_options[var]['shape']
-                control_units = phase.polynomial_control_options[var]['units']
-                shape = control_shape
-                units = control_units if con_units is None else con_units
-                linear = False
+                src = f'polynomial_control_values:{var}'
+                tgt = f'path_constraints.all_values:{con_name}'
 
             elif var_type == 'control_rate':
                 control_name = var[:-5]
-                constraint_path = 'control_rates:{0}_rate'.format(control_name)
-                phase.connect(src_name=constraint_path,
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                control_shape = phase.control_options[control_name]['shape']
-                control_units = phase.control_options[control_name]['units']
-                shape = control_shape
-                units = get_rate_units(control_units, time_units, deriv=1) \
-                    if con_units is None else con_units
-                linear = False
+                src = f'control_rates:{control_name}_rate'
+                tgt = f'path_constraints.all_values:{con_name}'
 
             elif var_type == 'control_rate2':
                 control_name = var[:-6]
-                constraint_path = 'control_rates:{0}_rate2'.format(control_name)
-                phase.connect(src_name=constraint_path,
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                control_shape = phase.control_options[control_name]['shape']
-                control_units = phase.control_options[control_name]['units']
-                shape = control_shape
-                units = get_rate_units(control_units, time_units, deriv=2) \
-                    if con_units is None else con_units
-                linear = False
+                src = f'control_rates:{control_name}_rate2'
+                tgt = f'path_constraints.all_values:{con_name}'
 
             elif var_type == 'polynomial_control_rate':
                 control_name = var[:-5]
-                constraint_path = 'polynomial_control_rates:{0}_rate'.format(control_name)
-                phase.connect(src_name=constraint_path,
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                control_shape = phase.polynomial_control_options[control_name]['shape']
-                control_units = phase.polynomial_control_options[control_name]['units']
-                shape = control_shape
-                units = get_rate_units(control_units, time_units, deriv=1) \
-                    if con_units is None else con_units
-                linear = False
+                src = f'polynomial_control_rates:{control_name}_rate'
+                tgt = f'path_constraints.all_values:{con_name}'
 
             elif var_type == 'polynomial_control_rate2':
                 control_name = var[:-6]
-                constraint_path = 'polynomial_control_rates:{0}_rate2'.format(control_name)
-                phase.connect(src_name=constraint_path,
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                control_shape = phase.polynomial_control_options[control_name]['shape']
-                control_units = phase.polynomial_control_options[control_name]['units']
-                shape = control_shape
-                units = get_rate_units(control_units, time_units, deriv=2) \
-                    if con_units is None else con_units
-                linear = False
+                src = f'polynomial_control_rates:{control_name}_rate2'
+                tgt = f'path_constraints.all_values:{con_name}'
 
             else:
                 # Failed to find variable, assume it is in the ODE
                 interleave_comp = phase._get_subsystem('interleave_comp')
-                phase.connect(src_name='interleave_comp.all_values:{0}'.format(con_name),
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
+                src = f'interleave_comp.all_values:{con_name}'
+                tgt = f'path_constraints.all_values:{con_name}'
 
                 ode_outputs = {opts['prom_name']: opts for (k, opts) in
                                phase.rhs_disc.get_io_metadata(iotypes=('output',)).items()}
@@ -389,7 +333,6 @@ class GaussLobatto(PseudospectralBase):
                 if var in ode_outputs:
                     shape = (1,) if len(ode_outputs[var]['shape']) == 1 else ode_outputs[var]['shape'][1:]
                     units = ode_outputs[var]['units'] if con_units is None else con_units
-                    linear = False
 
                     if interleave_comp.add_var(con_name, shape, units):
                         phase.connect(src_name='rhs_disc.{0}'.format(var),
@@ -401,14 +344,7 @@ class GaussLobatto(PseudospectralBase):
                                      f' the phase {phase.pathname} nor is it a known output of '
                                      f' the ODE.')
 
-            kwargs = options.copy()
-            kwargs.pop('shape', None)
-            kwargs.pop('units', None)
-            kwargs.pop('linear', None)
-            kwargs.pop('constraint_name', None)
-
-            path_comp._add_path_constraint_configure(con_name, shape=shape, units=units,
-                                                     linear=linear, **kwargs)
+            phase.connect(src_name=src, tgt_name=tgt)
 
     def configure_timeseries_outputs(self, phase):
         for timeseries_name, timeseries_options in phase._timeseries.items():
